lost_found/backend/requirements.txt
fastapi
uvicorn
streamlit
mysql-connector-python
python-multipart
pydantic
pymysql
python-dotenv
requests
Pillow